+++ CHANGELOG.md
@@ -20,7 +20,9 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * Application Request
       * Layout of the application board is destroyed for those applications which are fully "DONE" but application activation did not take place 
    * App Release Process
-      * validation fix for long description
+      * validation fix for long description in app page
+   * Service Release Process
+      * validation fix for long description in offer page
 
 * User Account
    * Add portal roles